chore: added redis and bullmq, removed uuid dependency ref: updated authGuard for checking account, reworked email confirmation entity, minor fixes and updates feat: added email confirmation part 2 (attempts included)

+++ backend/src/services/auth.service.ts
@@ -2,27 +2,33 @@ import type { IRefreshTokenRepository } from "../interfaces/i.refresh.token.repo
 import type { IUserRepository } from "../interfaces/i.user.repository.js"
 import refreshTokenRepository from "../repositories/refresh.token.repository.js"
 import userRepository from "../repositories/user.repository.js"
-import type {
-  UserRegisterDTO,
-  UserLoginDTO,
-  UserConfirmEmailDTO,
+import {
+  type UserRegisterDTO,
+  type UserLoginDTO,
+  type UserConfirmEmailDTO,
+  ErrorCode,
 } from "@project/shared"
 import type { AuthResponseDTO } from "../types/service.response.dto.js"
 import UserMapper from "../mappers/user.mapper.js"
 import { comparePassword, hashPassword } from "../utils/hash.js"
 import JWTHelper from "../utils/jwt.helper.js"
 import {
+  AppError,
   ConflictError,
   NotFoundError,
   UnauthorizedError,
+  ValidationError,
 } from "../errors/app.error.js"
 import type { UserCreateInput } from "../types/user.js"
 import emailSender from "../utils/email.sender.js"
-import type EmailSender from "../utils/email.sender.js"
+import EmailSender from "../utils/email.sender.js"
 import type { UserEntity } from "../domain/user.entity.js"
 import type { RefreshTokenEntity } from "../domain/refresh.token.entity.js"
 import type { IEmailConfirmationRepository } from "../interfaces/i.email.confirmation.repository.js"
 import emailConfirmationRepository from "../repositories/email.confirmation.repository.js"
+import { appConfig } from "../config/app.js"
+import prettifyTime from "../utils/prettify.time.js"
+import { randomInt } from "node:crypto"
 
 class AuthService {
   constructor(
@@ -34,7 +40,7 @@ class AuthService {
   async register(userData: UserRegisterDTO): Promise<AuthResponseDTO> {
     const exists = await this.userRepository.findByEmail(userData.email)
     if (exists) {
-      throw new ConflictError("User with this email already exists")
+      throw new AppError("User with this email already exists", 409, ErrorCode.ALREADY_EXISTS)
     }
     const hashedPassword = await hashPassword(userData.password)
     const createInput: UserCreateInput = {
@@ -42,28 +48,36 @@ class AuthService {
       name: userData.name,
       passwordHash: hashedPassword,
     }
-    const { user, refreshToken } = await this.userRepository.transaction<{
+    const { user, refreshToken, emailConfirmCode } = await this.userRepository.transaction<{
       user: UserEntity
       refreshToken: RefreshTokenEntity
+      emailConfirmCode: string
     }>(async (tx) => {
       const user = await this.userRepository.create(createInput, tx)
+      console.log(user)
       const refreshToken = await this.refreshTokenRepository.create(
         {
           sub: user.id,
+          expiresAt: new Date(Date.now() + appConfig.tokens.refresh.expire * 1000),
         },
         tx,
       )
-      const code = 1 + Math.floor(Math.random() * 899999 + 1) //create code and add to db
+      const code = randomInt(100000, 999999).toString()
       await this.emailConfirmationRepository.create(
-        { code, email: user.email },
+        {
+          code,
+          email: user.email,
+          expiresAt: new Date(Date.now() + appConfig.email.confirmExpire),
+        },
         tx,
       )
-      await this.emailSender.sendNotification("emailConfirm", {
-        target: user.email,
-        userName: user.name,
-        code,
-      })
-      return { user, refreshToken }
+      return { emailConfirmCode: code, user, refreshToken }
+    })
+    await this.emailSender.sendNotification("emailConfirm", {
+      target: user.email,
+      userName: user.name,
+      expiresIn: prettifyTime(appConfig.email.confirmExpire),
+      code: emailConfirmCode,
     })
     const { access, refresh } = JWTHelper.generateTokens({
       sub: user.id,
@@ -79,13 +93,11 @@ class AuthService {
   async login(userData: UserLoginDTO): Promise<AuthResponseDTO> {
     const user = await this.userRepository.findByEmail(userData.email)
     if (!user) throw new UnauthorizedError("Invalid credentials")
-    const isPasswordCorrect = await comparePassword(
-      userData.password,
-      user.passwordHash,
-    )
+    const isPasswordCorrect = await comparePassword(userData.password, user.passwordHash)
     if (!isPasswordCorrect) throw new UnauthorizedError("Invalid credentials")
     const dbToken = await this.refreshTokenRepository.create({
       sub: user.id,
+      expiresAt: new Date(Date.now() + appConfig.tokens.refresh.expire * 1000),
     })
     const { access, refresh } = JWTHelper.generateTokens({
       sub: user.id,
@@ -99,8 +111,7 @@ class AuthService {
     if (!decoded || !decoded.jti || !decoded.sub) throw new UnauthorizedError()
     const oldDbToken = await this.refreshTokenRepository.findByJti(decoded.jti)
     if (!oldDbToken) {
-      if (decoded.sub)
-        await this.refreshTokenRepository.deleteAllBySub(decoded.sub)
+      if (decoded.sub) await this.refreshTokenRepository.deleteAllBySub(decoded.sub)
       throw new UnauthorizedError()
     }
     const user = await this.userRepository.findById(oldDbToken.sub)
@@ -110,6 +121,7 @@ class AuthService {
     await this.refreshTokenRepository.delete(oldDbToken.jti)
     const newDbToken = await this.refreshTokenRepository.create({
       sub: oldDbToken.sub,
+      expiresAt: new Date(Date.now() + appConfig.tokens.refresh.expire * 1000),
     })
 
     const { access, refresh } = JWTHelper.generateTokens({
@@ -122,22 +134,27 @@ class AuthService {
   async confirmEmail(
     userData: { id: string; email: string },
     dto: UserConfirmEmailDTO,
-  ) {
-    const exists = await this.emailConfirmationRepository.findByEmail(
-      userData.email,
-    )
-    if (!exists) throw new NotFoundError("User")
+  ): Promise<void> {
+    const exists = await this.emailConfirmationRepository.findByEmail(userData.email)
+    if (!exists) throw new NotFoundError("Confirmation")
+
+    if (exists.expiresAt.getTime() <= new Date().getTime()) {
+      await this.emailConfirmationRepository.deleteByEmail(exists.email)
+      throw new AppError("Code expired", 410, ErrorCode.CODE_EXPIRED)
+    }
     if (exists.code !== dto.code) {
-      throw new ConflictError("Code does not match")
+      if (exists.attempts >= 2) {
+        await this.emailConfirmationRepository.deleteByEmail(exists.email)
+        throw new AppError("Out of attempts", 403, ErrorCode.TOO_MANY_ATTEMPTS)
+      }
+      await this.emailConfirmationRepository.incrementAttempt(exists.email)
+      throw new ValidationError([], "Code does not match")
     }
     await this.userRepository.transaction<void>(async (tx) => {
-      await this.emailConfirmationRepository.delete(exists.id, tx)
-      await this.userRepository.update(
-        userData.id,
-        { emailConfirmed: true },
-        tx,
-      )
+      await this.emailConfirmationRepository.deleteByEmail(exists.email, tx)
+      await this.userRepository.update(userData.id, { emailConfirmed: true }, tx)
     })
+    console.log(111)
   }
 }
